Cleanup: move "useCytoscapeConfig" and "useCyEffect" to hooks directory
